Rename utils.ts to promptUtils.ts to better reflect prompt handling usage
