+++ client/src/core/ai/promptUtils.ts
